Broke up arcane service into separate files
