+++ lib/arcane_framework.dart
@@ -40,7 +40,7 @@ library;
 export "package:arcane_framework/src/arcane.dart";
 export "package:arcane_framework/src/arcane_app.dart";
 export "package:arcane_framework/src/providers/environment_provider.dart";
-export "package:arcane_framework/src/providers/service_provider.dart";
+export "package:arcane_framework/src/providers/service/arcane_service.dart";
 export "package:arcane_framework/src/services/authentication/authentication_service.dart";
 export "package:arcane_framework/src/services/feature_flags/feature_flags_service.dart";
 export "package:arcane_framework/src/services/logging/logging_service.dart";
